% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-algebra.R
\name{-,ANY,LazyColumn-method}
\alias{-,ANY,LazyColumn-method}
\title{Vector Subtraction between regular vector and \code{LazyColumn}}
\usage{
\S4method{-}{ANY,LazyColumn}(e1, e2)
}
\arguments{
\item{e1}{A \code{LazyColumn} object.}

\item{e2}{A numeric vector.}
}
\value{
A numeric vector.
}
\description{
Subtracts a \code{LazyColumn} vector from a base vector.
}
\examples{
mat_a <- base::matrix(rnorm(12), nrow=3, ncol=4)
b <- rnorm(nrow(mat_a))
lazy_a <- LazyMatrix(mat_a, "sd", "mean")
lazy_col <- lazy_a[, 2]
b - lazy_col
}
