% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyMatrix-class.R
\name{as.matrix,LazyMatrix-method}
\alias{as.matrix,LazyMatrix-method}
\title{Attempts to turn the data matrix to a scaled matrix-object.}
\usage{
\S4method{as.matrix}{LazyMatrix}(x)
}
\arguments{
\item{x}{A LazyMatrix object.}
}
\value{
A matrix-object with scaled entries.
}
\description{
Attempts to turn the data matrix to a scaled matrix-object.
}
\examples{
mat_a <- base::matrix(rep(1, 6), nrow=2, ncol=3)
lazy_a <- LazyMatrix(mat_a, "sd", "mean")
as.matrix(lazy_a)
}
