% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{cache}
\alias{cache}
\title{Cache a reference object}
\usage{
cache(object)
}
\arguments{
\item{object}{Object to cache.}
}
\description{
This is a generic function used to retrieve data from a
reference object (such as a \code{\link[LazyMatrix]{LazyMatrix-class}}),
but the returned object remains a member of the reference class rather
than the base class.
}
\examples{
x = matrix(rnorm(1000 * 100), 1000, 100)
temp = tempfile(pattern = "example.matrix", fileext = ".tmp")
save(x, file = temp)
y = lazyMatrix(file = temp)

dim(y@.Data) # 1 x 1
z = cache(y)
dim(z@.Data) # 1000 x 100
class(z) # "LazyMatrix"

}
\seealso{
\code{\link[LazyMatrix]{instantiate}}
}
