% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{colnames,LazyMatrix-method}
\alias{colnames,LazyMatrix-method}
\title{Retrieve or set the row or column names of a LazyMatrix object.}
\usage{
\S4method{colnames}{LazyMatrix}(x)
}
\arguments{
\item{x}{A LazyMatrix object.}
}
\value{
A character vector of column names, or NULL if the matrix has no column names.
}
\description{
Retrieve or set the row or column names of a LazyMatrix object.
}
\examples{
mat_a <- base::matrix(rep(1, 6), nrow=2, ncol=3)
lazy_a <- LazyMatrix(mat_a, "sd", "mean")
colnames(lazy_a)
}
