% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{Compare,numeric,LazyColumn-method}
\alias{Compare,numeric,LazyColumn-method}
\title{Comparison operators for LazyColumn}
\usage{
\S4method{Compare}{numeric,LazyColumn}(e1, e2)
}
\arguments{
\item{e1}{A \code{LazyColumn} or numeric.}

\item{e2}{A \code{LazyColumn} or numeric.}
}
\value{
A logical vector.
}
\description{
Enables logical comparisons (\code{>}, \code{<}, \code{>=}, \code{<=}, \code{==}, \code{!=}) on a
\code{LazyColumn}, comparing against its scaled values.
}
