% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-algebra.R
\name{\%*\%,LazyColumn,LazyColumn-method}
\alias{\%*\%,LazyColumn,LazyColumn-method}
\title{Perform the dot product between two LazyColumn vectors}
\usage{
\S4method{\%*\%}{LazyColumn,LazyColumn}(x, y)
}
\arguments{
\item{x}{A \code{LazyColumn} object.}

\item{y}{A \code{LazyColumn} object.}
}
\value{
A numeric value with the resulting scalar.
}
\description{
Computes the inner product between two \code{LazyColumn} objects.
}
\examples{
mat_a <- base::matrix(rnorm(12), nrow=3, ncol=4)
X <- LazyMatrix(mat_a, "sd", "mean")
b <- rnorm(nrow(X))
lazy_col <- X[, 1]
lazy_col_2 <- X[, 2]
lazy_col \%*\% lazy_col_2
}
