% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-algebra.R
\name{\%*\%,LazyColumn,numeric-method}
\alias{\%*\%,LazyColumn,numeric-method}
\title{Perform the dot product between a LazyColumn and a numeric vector}
\usage{
\S4method{\%*\%}{LazyColumn,numeric}(x, y)
}
\arguments{
\item{x}{A \code{LazyColumn} object.}

\item{y}{A numeric scalar or vector.}
}
\value{
A numeric value with the resulting scalar.
}
\description{
Computes the inner product of a \code{LazyColumn} object
and a numeric vector.
If \code{y} is a scalar, scalar multiplication is performed.
If \code{y} is a vector of the same length as the column, the dot product
is performed.
}
\examples{
mat_a <- base::matrix(rnorm(12), nrow=3, ncol=4)
X <- LazyMatrix(mat_a, "sd", "mean")
b <- rnorm(nrow(X))
lazy_col <- X[, 1]
lazy_col \%*\% b
}
