% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-algebra.R
\name{\%*\%,numeric,LazyColumn-method}
\alias{\%*\%,numeric,LazyColumn-method}
\title{Perform the dot product between a numeric vectorand a LazyColumn}
\usage{
\S4method{\%*\%}{numeric,LazyColumn}(x, y)
}
\arguments{
\item{x}{A numeric scalar or vector.}

\item{y}{A \code{LazyColumn} object.}
}
\value{
A numeric value with the resulting scalar.
}
\description{
Computes the inner product of a numeric vector
and a \code{LazyColumn} object.
If \code{x} is a scalar, scalar multiplication is performed.
If \code{x} is a vector of the same length as the column, the dot product
is performed.
}
\examples{
mat_a <- base::matrix(rnorm(12), nrow=3, ncol=4)
X <- LazyMatrix(mat_a, "sd", "mean")
b <- rnorm(nrow(X))
lazy_col <- X[, 1]
b \%*\% lazy_col
}
