% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{instantiate}
\alias{instantiate}
\title{Instantiate a reference object}
\usage{
instantiate(object)
}
\arguments{
\item{object}{Object to instantiate.}
}
\description{
This is a generic function used to retrieve data from a
reference object (such as a \code{\link[LazyMatrix]{LazyMatrix-class}}) in
its base form (e.g., a \code{\link[base]{matrix}}).
}
\examples{
x = matrix(rnorm(1000 * 100), 1000, 100)
temp = tempfile(pattern = "example.matrix", fileext = ".tmp")
save(x, file = temp)
y = lazyMatrix(file = temp)

dim(y@.Data) # 1 x 1
z = instantiate(y)
dim(z@.Data) # 1000 x 100
class(z) # "matrix"

}
\seealso{
\code{\link[LazyMatrix]{cache}}
}
