% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lazy_crossprod_vec}
\alias{lazy_crossprod_vec}
\title{Fast crossprod for LazyMatrix (dense case)}
\usage{
lazy_crossprod_vec(x, s, c, y)
}
\arguments{
\item{x}{Dense matrix}

\item{s}{Column scale inverse vector (1 / col_scales)}

\item{c}{Column location vector (col_locations)}

\item{y}{Vector to multiply}
}
\value{
Vector result of t(X_tilde) * y
}
\description{
Fast crossprod for LazyMatrix (dense case)
}
