% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lazy_crossprod_vec_sp}
\alias{lazy_crossprod_vec_sp}
\title{Fast crossprod for LazyMatrix (sparse case)}
\usage{
lazy_crossprod_vec_sp(x, s, c, y)
}
\arguments{
\item{x}{Sparse matrix (dgCMatrix)}

\item{s}{Column scale inverse vector (1 / col_scales)}

\item{c}{Column location vector (col_locations)}

\item{y}{Vector to multiply}
}
\value{
Vector result of t(X_tilde) * y
}
\description{
Fast crossprod for LazyMatrix (sparse case)
}
