% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-definitions.R
\docType{class}
\name{LazyColumn-class}
\alias{LazyColumn-class}
\title{LazyColumn S4 class}
\value{
An object of class \code{LazyColumn} with slots \code{data} (numeric vector), \code{scale} (numeric scalar), and \code{location} (numeric scalar); represents a column of a \code{LazyMatrix} (scaled via scale and location).
}
\description{
An S4 class to represent a column vector as a subset of a LazyMatrix-object
}
\section{Slots}{

\describe{
\item{\code{data}}{The underlying data column vector.}

\item{\code{scale}}{Numeric scalar containing column-scale parameter.}

\item{\code{location}}{Numeric scalar containing the column-location parameter.}
}}

\examples{
mat <- matrix(1:6, nrow = 2, ncol = 3)
lazy_mat <- LazyMatrix(mat, "sd", "mean")
lazy_column <- lazy_mat[, 2]
}
