% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-definitions.R
\docType{class}
\name{LazyMatrix-class}
\alias{LazyMatrix-class}
\title{LazyMatrix S4 class}
\value{
An object of class \code{LazyMatrix} with slots \code{data} (matrix, possibly sparse), \code{col_scales}, \code{row_scales}, \code{col_locations}, \code{row_locations}.
Represents the original data matrix plus stored scaling/centering parameters used for lazy operations
}
\description{
An S4 class to represent a lazily transformed matrix with scaling and location parameters.
}
\section{Slots}{

\describe{
\item{\code{data}}{The underlying matrix.}

\item{\code{col_scales}}{Numeric vector of column scales.}

\item{\code{row_scales}}{Numeric vector of row scales.}

\item{\code{col_locations}}{Numeric vector of column locations.}

\item{\code{row_locations}}{Numeric vector of row locations.}
}}

\examples{
mat <- matrix(1:6, nrow=2, ncol=3)
obj <- LazyMatrix(mat, "sd", "mean")
}
