% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{length,LazyColumn-method}
\alias{length,LazyColumn-method}
\title{Get the length of a LazyColumn}
\usage{
\S4method{length}{LazyColumn}(x)
}
\arguments{
\item{x}{A LazyColumn object.}
}
\value{
An integer value containing the number of elements within the vector.
}
\description{
Get the length of a LazyColumn
}
\examples{
mat_a <- base::matrix(rep(1, 6), nrow=2, ncol=3)
lazy_a <- LazyMatrix(mat_a, "sd", "mean")
lazy_col <- lazy_a[, 2]
length(lazy_col)

}
