% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{lsqr}
\alias{lsqr}
\alias{lsqr,LazyMatrix-method}
\title{Performs least squares estimation on LazyMatrix object using the iterative lsqr algorithm.}
\usage{
lsqr(x, y, ...)

\S4method{lsqr}{LazyMatrix}(x, y)
}
\arguments{
\item{x}{A LazyMatrix object.}

\item{y}{A response vector.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A Matrix-object with the regression coefficients of the covariates.
}
\description{
Performs least squares estimation on LazyMatrix object using the iterative lsqr algorithm.
}
\examples{
set.seed(123)
mat_a <- base::matrix(rnorm(500), nrow = 50, ncol = 10)
lazy_a <- LazyMatrix(mat_a, "sd", "mean")
response_vector <- rnorm(nrow(mat_a))
lsqr(lazy_a, response_vector)
}
