% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyMatrix-class.R
\name{ncol,LazyMatrix-method}
\alias{ncol,LazyMatrix-method}
\title{Returns the number of columns of the data matrix}
\usage{
\S4method{ncol}{LazyMatrix}(x)
}
\arguments{
\item{x}{A LazyMatrix object.}
}
\value{
an integer of length 1 or NULL.
}
\description{
Returns the number of columns of the data matrix
}
\examples{
mat_a <- base::matrix(rep(1, 6), nrow=2, ncol=3)
lazy_a <- LazyMatrix(mat_a, "sd", "mean")
ncol(lazy_a)
}
