% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-algebra.R
\name{norm,LazyColumn-method}
\alias{norm,LazyColumn-method}
\title{Perform the norm of a LazyColumn vector}
\usage{
\S4method{norm}{LazyColumn}(x, type = "2")
}
\arguments{
\item{x}{A \code{LazyColumn} object.}

\item{type}{A character value defining type of norm. Default is Euclidean norm.}
}
\value{
A numeric value with the resulting scalar.
}
\description{
Computes the norm of a \code{LazyColumn} object.
}
\examples{
mat_a <- base::matrix(rnorm(12), nrow=3, ncol=4)
X <- LazyMatrix(mat_a, "sd", "mean")
b <- rnorm(nrow(X))
lazy_col <- X[, 1]
norm(lazy_col)
}
