% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-algebra.R
\name{+,LazyColumn,ANY-method}
\alias{+,LazyColumn,ANY-method}
\title{Vector Addition between \code{LazyColumn} and regular vector}
\usage{
\S4method{+}{LazyColumn,ANY}(e1, e2)
}
\arguments{
\item{e1}{A \code{LazyColumn} object.}

\item{e2}{A numeric vector.}
}
\value{
A numeric vector.
}
\description{
Sums a \code{LazyColumn} vector and a regular base vector.
}
\examples{
mat_a <- base::matrix(rnorm(12), nrow=3, ncol=4)
b <- rnorm(nrow(mat_a))
lazy_a <- LazyMatrix(mat_a, "sd", "mean")
lazy_col <- lazy_a[,2]
lazy_col + b
}
