% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-algebra.R
\name{+,LazyColumn,LazyColumn-method}
\alias{+,LazyColumn,LazyColumn-method}
\title{Vector Addition between two \code{LazyColumn} vectors.}
\usage{
\S4method{+}{LazyColumn,LazyColumn}(e1, e2)
}
\arguments{
\item{e1}{A \code{LazyColumn} object.}

\item{e2}{A \code{LazyColumn} object.}
}
\value{
A numeric vector.
}
\description{
Sums two \code{LazyColumn} vectors.
}
\examples{
mat_a <- base::matrix(rnorm(12), nrow=3, ncol=4)
lazy_a <- LazyMatrix(mat_a, "sd", "mean")
lazy_col <- lazy_a[,2]
lazy_col_2 <- lazy_a[, 3]
lazy_col_2 + lazy_col
}
