% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-algebra.R
\name{*,LazyColumn,LazyColumn-method}
\alias{*,LazyColumn,LazyColumn-method}
\title{Multiply element-wise two LazyColumn vectors}
\usage{
\S4method{*}{LazyColumn,LazyColumn}(e1, e2)
}
\arguments{
\item{e1}{A \code{LazyColumn} object.}

\item{e2}{A \code{LazyColumn} object.}
}
\value{
A numeric vector with the resulting vector.
}
\description{
Computes the element-wise multiplication of two \code{LazyColumn} objects
, preserving the lazy structure.
}
\examples{
mat_a <- base::matrix(rnorm(12), nrow=3, ncol=4)
X <- LazyMatrix(mat_a, "sd", "mean")
lazy_col <- X[, 1]
lazy_col_2 <- X[, 2]
lazy_col_2 * lazy_col
}
