% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-algebra.R
\name{*,LazyColumn,numeric-method}
\alias{*,LazyColumn,numeric-method}
\title{Multiply a LazyColumn by a numeric scalar or vector}
\usage{
\S4method{*}{LazyColumn,numeric}(e1, e2)
}
\arguments{
\item{e1}{A \code{LazyColumn} object.}

\item{e2}{A numeric scalar or vector.}
}
\value{
A numeric vector with the resulting vector.
}
\description{
Computes the element-wise multiplication of a \code{LazyColumn} object
with a numeric value, preserving the lazy structure.
If \code{e2} is a scalar, scalar multiplication is performed.
If \code{e2} is a vector of the same length as the column, element-wise
multiplication is performed.
}
\examples{
mat_a <- base::matrix(rnorm(12), nrow=3, ncol=4)
X <- LazyMatrix(mat_a, "sd", "mean")
lazy_col <- X[, 1]
lazy_col * 2
lazy_col * rnorm(nrow(X))
}
