% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-algebra.R
\name{\%*\%,LazyMatrix,ANY-method}
\alias{\%*\%,LazyMatrix,ANY-method}
\title{Matrix multiplication for \code{LazyMatrix} and vector}
\usage{
\S4method{\%*\%}{LazyMatrix,ANY}(x, y)
}
\arguments{
\item{x}{A \code{LazyMatrix} object.}

\item{y}{A numeric vector.}
}
\value{
A numeric matrix.
}
\description{
Multiplies a \code{LazyMatrix} object by a vector.
}
\examples{
mat_a <- base::matrix(rep(1, 6), nrow=2, ncol=3)
b <- c(1, 2, 3)
lazy_a <- LazyMatrix(mat_a, "sd", "mean")
lazy_a \%*\% b
}
