% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-algebra.R
\name{\%*\%,LazyMatrix,matrix-method}
\alias{\%*\%,LazyMatrix,matrix-method}
\title{Matrix multiplication for \code{LazyMatrix} and matrix-object.}
\usage{
\S4method{\%*\%}{LazyMatrix,matrix}(x, y)
}
\arguments{
\item{x}{A \code{LazyMatrix} object.}

\item{y}{A matrix-object.}
}
\value{
A matrix-object with the product of the lazy and non lazy object.
}
\description{
Multiplies a \code{LazyMatrix} object by a matrix
}
\examples{
mat_a <- matrix(rep(1, 6), nrow = 2, ncol = 3)
lazy_a <- LazyMatrix(mat_a, "sd", "mean")
set.seed(123)
m <- matrix(rnorm(6), nrow = 3, ncol = 2)
lazy_a \%*\% m
}
