% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyMatrix.R
\docType{class}
\name{LazyMatrix-class}
\alias{LazyMatrix-class}
\alias{lazyMatrix}
\title{An S4 class for lazy loading matrices}
\usage{
## Constructor:
lazyMatrix(..., name = NULL, package = NULL, file = NULL)
}
\description{
The LazyMatrix class extends \code{\link[base]{matrix}} by providing slots
that define a reference to a data matrix stored in a package or file.  Methods are
provided for most base functions to allow LazyMatrix objects to be used in place
of regular matrices; the file or package data is loaded on-the-fly when needed.  For other
applications (e.g., user-defined functions that expect a \code{matrix}), the
\code{\link[LazyMatrix]{instantiate}} method can be used to create a regular \code{matrix}
from a LazyMatrix.
}
\details{
LazyMatrix is a subclass of \code{\link[base]{matrix}} with three additional slots.
If these additional slots (name, package, and file) are NULL, a LazyMatrix behaves
as a regular matrix.  If either file or both name and package are not NULL, however, a LazyMatrix
behaves as if it contains the data in those locations (and ignores whatever is actually
stored in its @.Data slot, typically a 1 x 1 matrix with a value of NA).
}
\section{Slots}{

\describe{
\item{\code{...}}{Optional arguments to be passed to \code{\link[base]{matrix}}}

\item{\code{name}}{If the desired data resides in a package, the character string name of the data object}

\item{\code{package}}{If the desired data resides in a package, the character string name of the package}

\item{\code{file}}{If the desired data resides in a file, the character string full path of the file}
}}

\examples{
x = matrix(rnorm(1000 * 100), 1000, 100)
temp = tempfile(pattern = "example.matrix", fileext = ".tmp")
save(x, file = temp)
y = lazyMatrix(file = temp)

## Check dimensions
dim(x) # 1000 x 100
dim(y) # 1000 x 100

## Check object sizes
object.size(x) # ~800 KB
object.size(y) # ~1.5 KB

}
