% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyMatrix-class.R
\docType{class}
\name{LazyMatrix-class}
\alias{LazyMatrix-class}
\title{LazyMatrix S4 class}
\description{
An S4 class to represent a lazily transformed matrix with scaling and location parameters.
}
\section{Slots}{

\describe{
\item{\code{data}}{The underlying matrix.}

\item{\code{col_scales}}{Numeric vector of column scales.}

\item{\code{row_scales}}{Numeric vector of row scales.}

\item{\code{col_locations}}{Numeric vector of column locations.}

\item{\code{row_locations}}{Numeric vector of row locations.}
}}

\examples{
mat <- matrix(1:6, nrow=2, ncol=3)
obj <- LazyMatrix(mat, "sd", "mean")
}
