% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-definitions.R
\name{LazyMatrix}
\alias{LazyMatrix}
\title{Constructs a LazyMatrix object.}
\usage{
LazyMatrix(data, scale = NULL, location = NULL)
}
\arguments{
\item{data}{a matrix object.}

\item{scale}{optional scaling parameter.}

\item{location}{optional location parameter.}
}
\value{
A LazyMatrix object.
}
\description{
Constructs a LazyMatrix object.
}
\examples{
mat_a <- matrix(1:6, nrow=3, ncol=2)
lazy_a <- LazyMatrix(mat_a, scale="sd", location="mean")
lazy_a
}
