% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyMatrix-class.R
\name{lsqr,LazyMatrix-method}
\alias{lsqr,LazyMatrix-method}
\title{Performs least squares estimation on LazyMatrix object using the iterative lsqr algorithm.}
\usage{
\S4method{lsqr}{LazyMatrix}(x, y)
}
\arguments{
\item{x}{A LazyMatrix object.}

\item{y}{A response vector.}
}
\value{
A Matrix-object with the regression coefficients of the covariates.
}
\description{
Performs least squares estimation on LazyMatrix object using the iterative lsqr algorithm.
}
\examples{
mat_a <- base::matrix(rep(1, 6), nrow=2, ncol=3)
lazy_a <- LazyMatrix(mat_a, "sd", "mean")
reponse_vector <- c(1,2)
lsqr(lazy_a, response_vector)
}
