% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyMatrix-class.R
\name{norm}
\alias{norm}
\alias{norm,LazyMatrix-method}
\title{Computes the Frobenius norm of a LazyMatrix object.}
\usage{
norm(x)

\S4method{norm}{LazyMatrix}(x)
}
\arguments{
\item{x}{A LazyMatrix object.}
}
\value{
A numeric scalar representing the Frobenius norm of the matrix.
}
\description{
Computes the Frobenius norm of a LazyMatrix object.
}
\examples{
mat_a <- base::matrix(rnorm(50), nrow = 10, ncol = 5)
lazy_a <- LazyMatrix(mat_a, "sd", "mean")
norm(lazy_a)
}
