% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{nrow,LazyMatrix-method}
\alias{nrow,LazyMatrix-method}
\title{Returns the number of rows of the data matrix}
\usage{
\S4method{nrow}{LazyMatrix}(x)
}
\arguments{
\item{x}{A LazyMatrix object.}
}
\value{
an integer of length 1 or NULL.
}
\description{
Returns the number of rows of the data matrix
}
\examples{
mat_a <- base::matrix(rep(1, 6), nrow=2, ncol=3)
lazy_a <- LazyMatrix(mat_a, "sd", "mean")
nrow(lazy_a)
}
