% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-algebra.R
\name{svd,LazyMatrix-method}
\alias{svd,LazyMatrix-method}
\title{Singular Value decomposition for LazyMatrix.}
\usage{
\S4method{svd}{LazyMatrix}(x, nu = min(n, p), nv = min(n, p))
}
\arguments{
\item{x}{A LazyMatrix object.}

\item{nu}{number of left singular vectors to estimate (defaults to nv).}

\item{nv}{number of right singular vectors to estimate.}
}
\value{
A list with entries:
\item{d}{max(nu, nv) approximate singular values}
\item{u}{nu approximate left singular vectors (only when right_only=FALSE)}
\item{v}{nv approximate right singular vectors}
\item{iter}{The number of Lanczos iterations carried out}
\item{mprod}{The total number of matrix vector products carried out}
}
\description{
Performs lazy SVD using irlba for partial Singular value decomposition on sparse matrices.
}
\examples{
set.seed(123)
mat_a <- matrix(rnorm(500), nrow = 50, ncol = 10)
lazy_a <-LazyMatrix(mat_a, scale = "sd", location = "mean")
S <- svd(lazy_a)
# Receive singular values with
S$d
}
