% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-algebra.R
\name{t,LazyMatrix-method}
\alias{t,LazyMatrix-method}
\title{Given a \code{LazyMatrix} x, t returns the transpose of x.}
\usage{
\S4method{t}{LazyMatrix}(x)
}
\arguments{
\item{x}{A \code{LazyMatrix} object.}
}
\value{
A \code{LazyMatrix} object with the transposed data matrix.
}
\description{
Given a \code{LazyMatrix} x, t returns the transpose of x.
}
\examples{
mat_a <- base::matrix(rep(1, 6), nrow=2, ncol=3)
lazy_a <- LazyMatrix(mat_a, "sd", "mean")
lazy_t <- t(lazy_a)
}
